requirements.txt
# Core web interface
gradio>=5.0.0
gradio-client>=1.10.0
fastapi>=0.115.0
uvicorn>=0.24.0
aiofiles>=22.0.0
python-multipart>=0.0.9

# AI/ML APIs
anthropic>=0.57.0
openai>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tenacity>=8.0.0
python-dateutil>=2.8.0

# Search and retrieval
rank-bm25>=0.2.2
pinecone>=7.0.0
chromadb>=0.4.0
scikit-learn>=1.3.0

# AWS cloud storage dependencies
boto3>=1.34.0
psycopg[binary]>=3.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# Additional UI dependencies
pillow>=8.0.0
orjson>=3.0.0

# Optional dependencies for full functionality
# alembic>=1.12.0  # For database migrations
